+++ src/Main/MainScreen.java
@@ -3,6 +3,7 @@ package Main;
 import AdvanceListener.Advance;
 import Populate.Planet;
 import Populate.ShowVisibleVerify;
+import Starship.Starship;
 import javafx.application.Application;
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
@@ -61,23 +62,6 @@ public class MainScreen extends Application
     AtomicInteger distance = new AtomicInteger(0);
     String unitsFromEarth = "light years";
     String errorMessage = "-ERROR, YOU'VE GOT A DARN BUG IN YOUR CODE!";
-    String[] shipEngine = {"Gen II warp-drive", "Gen III warp-drive", "Gen IV" +
-            " warp-drive", "Gen V warp-drive", "Dark matter propulsion drive"
-            , "Wormhole drive", "Blackhole warmhole engine"};
-    int shipEngineCurrentNum = 0;
-    int engineSpeed = 1; // light years
-    String[] shipHull = {"Reinforced steel", "Double-reinforced steel",
-            "Reinforced titanium", "Double-reinforced titanium"};
-    int shipHullCurrentNum = 0;
-    int shipHullPoints = 1;
-    String[] shipWeapons = {"Basic ship rounds", "High-caliber ship rounds",
-            "Modern ship rounds", "Advanced ship rounds", "Crude phasers",
-            "Low-powered phasers", "High-powered phasers", "Dark-matter " +
-            "phasers"};
-    int shipWeaponsCurrentNum = 0;
-    int shipWeaponPoints = 1;
-    double shipFuelStorage = 100.0;
-    double shipFuel = 100.0;
     Planet currentPlanet = null;
     GridPane outOfFuelGridPane;
     AtomicBoolean occupiedSpace = new AtomicBoolean(false);
@@ -93,7 +77,7 @@ public class MainScreen extends Application
     Text mineralOutputText = new Text();
     Text scienceOutputText = new Text();
     ArrayList<Planet> populatedPlanetArrayList = new ArrayList<>();
-
+    Starship starship = new Starship();
 
 
 /*    public void restart()
@@ -136,11 +120,13 @@ public class MainScreen extends Application
             Scene scene = new Scene(borderPane);
             stage.setScene(scene);
             stage.show();
-            ShopScene shopScene = new ShopScene(shipEngine, engineSpeed, shipHull,
-                    shipHullPoints, shipWeapons, shipWeaponPoints, stage,
-                    scene, populatedPlanetArrayList, shipEngineCurrentNum,
-                    shipHullCurrentNum, shipWeaponsCurrentNum,
-                    shipFuelStorage);
+            ShopScene shopScene = new ShopScene(starship.getShipEngine(),
+                    starship.getEngineSpeed(), starship.getShipHull(),
+                    starship.getShipHullPoints(), starship.getShipWeapons(),
+                    starship.getShipWeaponPoints(), stage, scene,
+                    populatedPlanetArrayList, starship.getShipEngineCurrentNum(),
+                    starship.getShipHullCurrentNum(), starship.getShipWeaponsCurrentNum(),
+                    starship.getShipFuelStorage());
             PlanetsScene planetsScene = new PlanetsScene(stage, scene,
                     shopScene, populatedPlanetArrayList);
 
@@ -183,13 +169,16 @@ public class MainScreen extends Application
             Line newLine = new Line();
             newLine.setStartX(260);
             newLine.setEndX(575);
-            Text shipEngineText = new Text("Ship engine: " + shipEngine[0]);
-            Text shipHullText = new Text("Ship hull: " + shipHull[0]);
-            Text shipWeaponsText = new Text("Ship weapons: " + shipWeapons[0]);
+            Text shipEngineText =
+                    new Text("Ship engine: " + starship.getShipEngine()[0]);
+            Text shipHullText =
+                    new Text("Ship hull: " + starship.getShipHull()[0]);
+            Text shipWeaponsText =
+                    new Text("Ship weapons: " + starship.getShipWeapons()[0]);
             Text shipFuelStorageText =
-                    new Text("Ship fuel storage capacity: " + shipFuelStorage);
+                    new Text("Ship fuel storage capacity: " + starship.getShipFuelStorage());
             Text shipFuelText =
-                    new Text("Fuel: " + shipFuel);
+                    new Text("Fuel: " + starship.getShipFuel());
 
 
             Label distanceText =
@@ -249,8 +238,8 @@ public class MainScreen extends Application
             borderPane.setRight(infoBox);
 
             // Advance button handler
-            Advance advance = new Advance(distance, engineSpeed,
-                    this);
+            Advance advance = new Advance(distance, starship.getEngineSpeed(),
+                    this, starship);
 
 
             // Pop-ups *section*
@@ -521,7 +510,7 @@ public class MainScreen extends Application
                         alreadyPopulatedText.setVisible(false);
                         // Update the distance label
                         distanceText.setText(distance.get() + " " + unitsFromEarth + " from Earth");
-                        shipFuelText.setText("Fuel: " + shipFuel);
+                        shipFuelText.setText("Fuel: " + starship.getShipFuel());
 
 
                         // Update the infoText label
@@ -537,7 +526,7 @@ public class MainScreen extends Application
                             material.setText("Material on planet: " + currentPlanet.getMaterials());
                             planetType.setText("Planet type: " + currentPlanet.getPlanetType());
                         }
-                        if (shipFuel <= 0 && currentPlanet == null)
+                        if (starship.getShipFuel() <= 0 && currentPlanet == null)
                         {
                             outOfFuelGridPane.setVisible(true);
                             occupiedSpace.set(true);
@@ -682,108 +671,8 @@ public class MainScreen extends Application
             this.currentPlanet = newPlanet;  // Update the current planet
         }
 
-        public double getShipFuel()
-        {
-            return shipFuel;
-        }
-
-        public void setShipFuel(double newFuel)
-        {
-            shipFuel = newFuel;
-        }
-
         public GridPane getOutOfFuelGridPane()
         {
             return outOfFuelGridPane;
         }
-
-    public String[] getShipEngine()
-    {
-        return shipEngine;
-    }
-
-    public void setShipEngine(String[] shipEngine)
-    {
-        this.shipEngine = shipEngine;
-    }
-
-    public int getShipEngineCurrentNum()
-    {
-        return shipEngineCurrentNum;
-    }
-
-    public void setShipEngineCurrentNum(int shipEngineCurrentNum)
-    {
-        this.shipEngineCurrentNum = shipEngineCurrentNum;
-    }
-
-    public int getEngineSpeed()
-    {
-        return engineSpeed;
-    }
-
-    public void setEngineSpeed(int engineSpeed)
-    {
-        this.engineSpeed = engineSpeed;
-    }
-
-    public String[] getShipHull()
-    {
-        return shipHull;
-    }
-
-    public void setShipHull(String[] shipHull)
-    {
-        this.shipHull = shipHull;
-    }
-
-    public int getShipHullCurrentNum()
-    {
-        return shipHullCurrentNum;
-    }
-
-    public void setShipHullCurrentNum(int shipHullCurrentNum)
-    {
-        this.shipHullCurrentNum = shipHullCurrentNum;
-    }
-
-    public int getShipHullPoints()
-    {
-        return shipHullPoints;
-    }
-
-    public void setShipHullPoints(int shipHullPoints)
-    {
-        this.shipHullPoints = shipHullPoints;
-    }
-
-    public String[] getShipWeapons()
-    {
-        return shipWeapons;
-    }
-
-    public void setShipWeapons(String[] shipWeapons)
-    {
-        this.shipWeapons = shipWeapons;
-    }
-
-    public int getShipWeaponsCurrentNum()
-    {
-        return shipWeaponsCurrentNum;
-    }
-
-    public void setShipWeaponsCurrentNum(int shipWeaponsCurrentNum)
-    {
-        this.shipWeaponsCurrentNum = shipWeaponsCurrentNum;
-    }
-
-    public int getShipWeaponPoints()
-    {
-        return shipWeaponPoints;
-    }
-
-    public void setShipWeaponPoints(int shipWeaponPoints)
-    {
-        this.shipWeaponPoints = shipWeaponPoints;
-    }
 }
